@@ -1,14 +1,19 @@
 require 'rest-client'
 require 'pry'
 
-SCHEDULER.every '10m', :first_in => 0 do |job|
+SCHEDULER.every '60s', :first_in => 60 do |job|
   atlas_stats = project_status ENV["CODE_CLIMATE_ATLAS_ID"]
-  # leitor10_stats = project_status ENV["CODE_CLIMATE_LEITOR10_ID"]
+  leitor10_stats = project_status ENV["CODE_CLIMATE_LEITOR10_ID"]
+  eva_stats = project_status ENV["CODE_CLIMATE_EVA_ID"]
+
   send_event("atlas_codeclimate", atlas_stats)
   send_event("atlas_coverage", atlas_stats)
 
-  # send_event("leitor10_codeclimate", leitor10_stats)
-  # send_event("leitor10_coverage", leitor10_stats)
+  send_event("leitor10_codeclimate", leitor10_stats)
+  send_event("leitor10_coverage", leitor10_stats)
+
+  send_event("eva_codeclimate", eva_stats)
+  send_event("eva_coverage", eva_stats)
 end
 
 def project_status repo_id
