@@ -5,9 +5,10 @@ require 'pry'
 SCHEDULER.every '2m', :first_in => 0 do |job|
   send_event('atlas_last_deploy', last_deploy('Atlas', ENV["ROLLBAR_ACCESS_TOKEN_ATLAS"]))
   send_event('leitor10_last_deploy', last_deploy('Leitor 10', ENV["ROLLBAR_ACCESS_TOKEN_LEITOR10"]))
+  send_event('eva_last_deploy', last_deploy('EVA', ENV["ROLLBAR_ACCESS_TOKEN_EVA"]))
 end
 
-def last_deploy project_name, access_token  
+def last_deploy project_name, access_token
   last_deploy = nil
   count = 1
   while !last_deploy
